+++ lamindb/_save.py
@@ -174,14 +174,14 @@ def upload_data_object(file) -> None:
     # do NOT hand-craft the storage key!
     file_storage_key = auto_storage_key_from_file(file)
     if hasattr(file, "_to_store") and file._to_store and file.suffix != ".zarr":
-        logger.hint(f"storing object {file.name} with key {file_storage_key}")
+        logger.hint(f"storing file {file.name} with key {file_storage_key}")
         store_object(file._local_filepath, file_storage_key)
     elif (
         file.suffix == ".zarr"
         and hasattr(file, "_memory_rep")
         and file._memory_rep is not None
     ):
-        logger.hint(f"storing object {file.name} with key {file_storage_key}")
+        logger.hint(f"storing file {file.name} with key {file_storage_key}")
         storagepath = lamindb_setup.settings.storage.key_to_filepath(file_storage_key)
         print_progress = partial(print_hook, filepath=file.name)
         write_adata_zarr(file._memory_rep, storagepath, callback=print_progress)