+++ lamindb/dev/__init__.py
@@ -8,9 +8,11 @@
    datasets
    hashing
    storage
+   Settings
 """
 
 from lnschema_core._queryset import QuerySet
 from lnschema_core.models import BaseORM
 
 from . import datasets  # noqa
+from ._settings import Settings